[PATCH] ppc64: Janitor rtas_call() return variables
Recently I changed the return value of rtas_call() from an unsigned long to
an int.  That patch missed a few places where we declare a variable to
store the result from rtas_call().  This new patch changes those places to
use an int variable instead of a long or unsigned long variable.  Linas
Vepstas pointed this out.

+++ arch/ppc64/kernel/lparcfg.c
@@ -245,7 +245,7 @@ unsigned int get_splpar_potential_characteristics(void)
 	/* return 0 for now.  Underlying rtas functionality is not yet complete. 12/01/2003*/
 	return 0; 
 #if 0 
-	long call_status;
+	int call_status;
 	unsigned long ret[2];
 
 	char * buffer = kmalloc(SPLPAR_MAXLENGTH, GFP_KERNEL);
